requirements.txt
asgiref==3.3.0
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
Markdown==3.3.3
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
coreapi==2.3.3
